Add files via upload
RFID_aggregate.5.3 performs the same function as previous versions, but saves updated spreadsheets for individual fishids by month (the Fishid itself is a folder containing .xlsx files for each month for which there was an RFID reading for that fishid). aggregate_export(), aggregate_raster(), and aggregate_barplot() still need to be updated to read in these split-by-month files.